Added EventHandlerResult type
a la Kaleidoscope, this is an enum that gives more possibility for messages than the
boolean currently used as the return value for plugin event handlers. I think I might
change it to `PluginHookResult` or something like that, and maybe change it from an enum
to a bit field, but for now, I'll stil to mirroring Kaleidoscope. The big difference so
far is the `nxplugin` value, which will be used in `handleKeyEvent` in case it calls the
hook function for a plugin that doesn't implement the event handler hook.