Add solid-color LED background mode
This also changes the semantics of how `LedController` deals with background modes. If the
current mode index is invalid, it treats that as "off" for background LEDs. No more checks
are done for `nullptr`, and when cycling through modes with `nextMode()`, we let it run
off the end of the array, to have a default "off" mode without requiring any defined mode
for it.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>