Move check for toggled on active key
This change moves the check for a non-empty entry in `active_keys_[]` when a key toggles
on from before plugin hooks are called to after. Now we wait until just before setting the
entry in `active_keys_` from `event.key`, so plugins such as Glukeys don't have to deal
with premature release events.

Note that this approach has its own drawbacks, because we now send a release event,
calling all plugin hooks again, after they think an event has been processed, but before
its report gets sent to the host. This is probably not a big deal because this is really a
just-in-case measure, anyway, but maybe we can do better.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>