Rewrite LED mode management
This is a big change that allows LED modes to share space in RAM for the parts
that don't need to persist when a given mode is not the active one. It took me a
long time to get this to compile, so I'm checking it in despite the fact that it
doesn't work properly yet.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>